Refactoring content of application/accounts folder to two folders: application/accounts and application/users
