+++ backend/Application.UnitTests/Common/Mappings/MappingTests.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using Domain.Entities;
 using System;
 using Application.Accounts;
+using Application.Users;
 using Xunit;
 
 namespace Application.UnitTests.Common.Mappings
