+++ backend/Domain/Entities/IUser.cs
@@ -10,7 +10,7 @@ namespace Domain.Entities
     public int Id { get; set; }
     public string Email { get; set; }
     public string Password { get; set; }
-    public RoleEnum Role { get; set; }
+    public RoleEnum Role { get; }
     public string Name { get; set; }
     public DateTimeOffset? DeactivationTime { get; set; }
   }
