+++ backend/Web/Controllers/AuthController.cs
@@ -1,9 +1,8 @@
 using Microsoft.AspNetCore.Mvc;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using Application.Accounts;
-using Application.Accounts.Commands.CheckAuthCommand;
-using Application.Accounts.Commands.CreateAccountCommand;
+using Application.Users;
+using Application.Users.Commands.CheckAuthCommand;
+using Application.Users.Commands.CreateAccountCommand;
 
 namespace Web.Controllers
 {
